1) Various fixes for compatibility with msvc 2) Modification to the progress report string 3) Fixed PrecurorWaterLoss and PrecursorDoubleWaterLoss erasing code in TagreconSpectrum.cpp
Former-commit-id: 315fab98fe6ae3f466ef1fdc263da4f368ef8319

+++ pwiz_tools/Bumbershoot/tagrecon/tagrecon.cpp
@@ -1394,7 +1394,7 @@ namespace tagrecon
 
 					simplethread_lock_mutex( &resourceMutex );
 					cout << threadInfo->workerHostString << " has searched " << threadInfo->stats.numProteinsDigested << " of " <<	numProteins <<
-							" proteins; " << proteinsPerSec << " per second, " << totalSearchTime << " elapsed, " << estimatedTimeRemaining << " remaining." << endl;
+							" proteins; " << ((int) proteinsPerSec) << " per second, " << ((int)totalSearchTime) << " elapsed, " << ((int) estimatedTimeRemaining) << " remaining." << endl;
 
 					PRINT_PROFILERS(cout, threadInfo->workerHostString + " profiling")
